refactor: Use float64 for position to improve movement smoothness feat: Add GitHub Actions workflow for running tests
- Changed Position struct and related logic to use float64.
- Updated tests to reflect float64 usage.
- Added basic GitHub Actions workflow to run 'make test' on push/pull_request.

+++ types.go
@@ -20,10 +20,10 @@ const (
 	// Add more types if needed
 )
 
-// Position represents the X, Y coordinates on the terminal screen.
+// Position represents the X, Y coordinates on the terminal screen using float64 for smoother movement.
 type Position struct {
-	X int
-	Y int
+	X float64
+	Y float64
 }
 
 // Runner represents a single animated runner on the screen.
